Move helpers into EnrollCore component
