+++ components/enroll_core/app/helpers/enroll_core/brokers/welcome_helper.rb
