+++ components/enroll_core/app/helpers/enroll_core/events_helper.rb
