+++ components/enroll_core/app/helpers/enroll_core/insured/group_selection_helper.rb
