+++ components/enroll_core/app/helpers/enroll_core/notices_helper.rb
