+++ components/enroll_core/app/helpers/enroll_core/permission_helper.rb
