+++ components/enroll_core/app/helpers/enroll_core/portal_header_helper.rb
