+++ components/enroll_core/lib/set_current_user.rb
