+++ components/enroll_core/spec/lib/set_current_user_spec.rb
